feat: Initial, unrefined solution
My first working solution. As is, this will simply prevent displaying
any "Required material types gathered" wishlist notifications, i.e. the
ones that simply mean that you have enough category materials. This
solution is a bit naive, as while it is unlikely, it is possible to have
enough of the primary materials but not enough category materials, which
would still be misleading.